Change precedence order when searching the KeyCode
Remove event.code from the expression that tries to find out what the
key code for a pressed key was.

During initial testing this key was not being returned by Chrome, so
everything worked, but "suddenly" it started to have a value, which made
the shortkeys to stop being recognized.

In theory we should receive the key code in key attribute, but Chrome
does not conform to the Web Standards and therefore we need to check
which and keyCode and keyChar for the value.