Add notebook navigation
This patch adds notebook navigation using j, k and notebook selection
using enter or l keys.

It works with or without a search filter.

Now if you do a search and hit the enter key you will go directly to
navigating the results, but if there's only 1 result it will select
it.