Add search scope selecting and search saving
We add 2 more shortcuts in the Note search:

w - Select the scope of the search (where to search)
a - Archive/Save a search for later use

There a bug in Evernote that sometimes hide the saving button even when
it should be displayed. The shortcut will work even when the search is
not being displayed.