Fix bug in change note title/caption
This patch fixes a bug where you couldn't change a note's title/caption
if it was not visible even when you where in a note view.

This happened for example when you were in the middle of a note and you
wanted to change the title.

To fix this we now support a new feature in the keymanager, we can now
set in visible a string with a path of an alternative element to check
for visibility if the element itself is not visible.  Either one's
visibility pass the visibility check.