Fix a problem navigating the notebooks.
This patch fixes a couple of problems navigating notebooks.

The way notebooks were selected depended on CSS class naming which
changed so everything stopped working, so we need a different way to
detect which notebook was selected.

Also if the notebook search field is empty and we press Enter it would
select one notebook regardless.