Fix exit note editing with ESC
Exiting from editing a note with ESC key stopped working because we were
searching for an ID that is dinamically generated by Evernotes, so when
the ID changes the code stops finding the iframe.

This pach changes the way the observer searches for the iframe, now we
search for the name field which is less likely to change.