Fix saving search shortcut
Saving a search used the class to search for the "Save search" button,
but this class is dynamically generated by Evernote, so as soon as it
changes the functionality broke.

This patch changes this by transversing the document and search for the
"Save search" content if we are in the search page and we have contents
in the search field.

Even though searching the document is not optimal, since the button has
no ID, there is no better way at the moment.

We don't search the whole document, we just start at the closest parent
that has an ID.