Fix shortcuts on focused note
When a note was focused some shortcuts were still enabled, like entering
the search field, clearing the seach, saving the search, etc., and they
shouldn't since we don't see what the actions are doing and in some
cases we even go into editing mode in a field that is hidden.

This patch adds visibility constraints to those actions so they are only
executed when it makes sense.