Fix issues with the tag search field
We had problems when entering and exiting the search tag field.

The problem on entering was that the class being used to locate the
field no longer existend, replaced with the input class.

The problem on exit was that the tab key code we generate to exit the
field would bring us to the browser's address bar.

We replace the exit of all search fields with Shift + Tab instead of
just Tab.