Fix bug in change note title on new note
When we are creating a new note if we exit the title before entering any
text we will not be able to enter it with the "c" key unless we enter
some text in the contents section.

This patch fixes this by changing required visibility to the notebook
select menu id.