Fix problem with keycode recognition
In some cases keys where not recognized if Chrome key event passed a
value in key attribute.

This patch changes the priority of the attributes and also makes sure
that we are always comparing key codes even if we get a char in the
event so that they can be correctly matched.