Fix field exit with ESC
Pressing ESC no longer exits from test fields:
- Note title
- Workchat search
- Tag search
- Notebook search
- Search

The reason is that Chrome no longer invoke default actions when
dispatching synthetic events, so the Shift+TAB key that we were
simulating no longer works.

The new mechanism used to fix this is to create an invisible element,
give it focus, and then remove it.