Add more tooltips shortcut indicators
With this patch we add shortcut indicators to more tooltips via a DOM
mutation observer that dynamically modifies the contents of the div.

Ideally we would be modifying the source string, but unfortunately this
data is retrieved and stored as part of a closure, so even if we modify
the source string, for example in
`window.gwtExport.FocusMessages.tooltip_Format_Drive` it is no longer
accessed afterwards.

The "Create a tag" tooltip doesn't get modified because it modifies the
"span" instead of using the same mechanism as any of the other tooltips.