Fix notebook movements
Upwards and downwards notebook movements are broken due to a change on
the alpha channel of the selected notebook color.

We are using the color to detect which is the currently selected
notebook, so the change of any of its parameters prevents us from
properly detecting it.

This patch starts ignoring the alpha channel on the detection, which
fixes the issue and hopefully prevents future bugs if they decide to
change the alpha channel again.